IVF-main/backend/requirements.txt
aniso8601==10.0.1
annotated-types==0.7.0
bcrypt==4.1.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
chardet==6.0.0.post1
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Cors==3.0.10
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-ai-generativelanguage==0.6.6
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.7.2
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.62.3
httplib2==0.31.2
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.2
openpyxl==3.1.5
pandas==2.2.2
passlib==1.7.4
pillow==12.1.1
proto-plus==1.27.1
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
reportlab==4.2.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.31
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.0.3
WTForms==3.1.2
